move _build_worker out of the server module to the test_listen module
